refactor(cvInputs): move CV input handling to new cvInputs.hpp file
