+++ firmware-CLK/lib/cvInputs.hpp
@@ -1,0 +1,269 @@
+#pragma once
+
+// cvInputs.hpp — CV input processing, CVTarget enum, attenuation, calibration
+//
+// Owns: CVTarget enum, CVTargetDescription[], CVInputTarget/Attenuation/Offset[],
+//       channelADC[], offsetScale[], pendingCVInputTarget[],
+//       AdjustADCReadings(), HandleCVInputs(), HandleCVTarget().
+// Depends on: outputs[], BPM/tickCounter/externalTickCounter (clockEngine.hpp),
+//             masterState/SetMasterState() (main.cpp via extern).
+
+#include <Arduino.h>
+
+#include "boardIO.hpp"
+#include "clockEngine.hpp"
+#include "outputs.hpp"
+#include "pinouts.hpp"
+#include "utils.hpp"
+
+// ── CV modulation target enum ─────────────────────────────────────────────────
+enum CVTarget {
+    None = 0,
+    StartStop,
+    Reset,
+    SetBPM,
+    Div1,
+    Div2,
+    Div3,
+    Div4,
+    Output1Prob,
+    Output2Prob,
+    Output3Prob,
+    Output4Prob,
+    Swing1Amount,
+    Swing1Every,
+    Swing2Amount,
+    Swing2Every,
+    Swing3Amount,
+    Swing3Every,
+    Swing4Amount,
+    Swing4Every,
+    Output3Level,
+    Output4Level,
+    Output3Offset,
+    Output4Offset,
+    Output3Waveform,
+    Output4Waveform,
+    Output1Duty,
+    Output2Duty,
+    Output3Duty,
+    Output4Duty,
+    Envelope1,
+    Envelope2,
+    Output3,
+    Output4,
+};
+
+String CVTargetDescription[] = {
+    "None",
+    "Start/Stop",
+    "Reset",
+    "Set BPM",
+    "Output 1 Div",
+    "Output 2 Div",
+    "Output 3 Div",
+    "Output 4 Div",
+    "Output 1 Prob",
+    "Output 2 Prob",
+    "Output 3 Prob",
+    "Output 4 Prob",
+    "Swing 1 Amt",
+    "Swing 1 Every",
+    "Swing 2 Amt",
+    "Swing 2 Every",
+    "Swing 3 Amt",
+    "Swing 3 Every",
+    "Swing 4 Amt",
+    "Swing 4 Every",
+    "Output 3 Lvl",
+    "Output 4 Lvl",
+    "Output 3 Off",
+    "Output 4 Off",
+    "Output 3 Wav",
+    "Output 4 Wav",
+    "Output 1 Duty",
+    "Output 2 Duty",
+    "Output 3 Duty",
+    "Output 4 Duty",
+    "Output 3 Env",
+    "Output 4 Env",
+    "Output 3",
+    "Output 4",
+};
+int CVTargetLength = sizeof(CVTargetDescription) / sizeof(CVTargetDescription[0]);
+
+// ── CV input state globals ────────────────────────────────────────────────────
+CVTarget pendingCVInputTarget[NUM_CV_INS] = {CVTarget::None, CVTarget::None};
+
+// ADC input offset and scale from calibration
+float offsetScale[NUM_CV_INS][2]; // [channel][0: offset, 1: scale]
+
+// Active CV target assignments
+CVTarget CVInputTarget[NUM_CV_INS]  = {CVTarget::None, CVTarget::None};
+int      CVInputAttenuation[NUM_CV_INS] = {0, 0};
+int      CVInputOffset[NUM_CV_INS]      = {0, 0};
+
+// ADC readings (calibrated, filtered)
+float channelADC[NUM_CV_INS], oldChannelADC[NUM_CV_INS];
+
+// ── extern refs defined in main.cpp / clockEngine.hpp ────────────────────────
+extern float    ADCCal[2];
+extern int      ADCOffset[2];
+extern bool     masterState;
+extern void     SetMasterState(bool state);
+
+// Forward declaration
+void HandleCVTarget(int ch, float CVValue, CVTarget cvTarget);
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Read one ADC channel and apply calibration into channelADC[ch]
+// ─────────────────────────────────────────────────────────────────────────────
+void AdjustADCReadings(int CV_IN_PIN, int ch) {
+    float calibratedReading = max((analogRead(CV_IN_PIN) - ADCOffset[ch]) * ADCCal[ch], 0.0f);
+    channelADC[ch] = calibratedReading;
+}
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Poll both CV inputs and dispatch to HandleCVTarget() on meaningful change
+// ─────────────────────────────────────────────────────────────────────────────
+void HandleCVInputs() {
+    for (int i = 0; i < NUM_CV_INS; i++) {
+        oldChannelADC[i] = channelADC[i];
+        AdjustADCReadings(CV_IN_PINS[i], i);
+        ONE_POLE(channelADC[i], oldChannelADC[i], 0.5f);
+        if (abs(channelADC[i] - oldChannelADC[i]) > 10) {
+            HandleCVTarget(i, channelADC[i], CVInputTarget[i]);
+        }
+    }
+}
+
+// ─────────────────────────────────────────────────────────────────────────────
+// Apply a CV value to its assigned target parameter
+// ─────────────────────────────────────────────────────────────────────────────
+volatile bool lastResetState = false;
+
+void HandleCVTarget(int ch, float CVValue, CVTarget cvTarget) {
+    // Attenuate and offset the CVValue
+    float attenuatedValue = CVValue * ((100 - CVInputAttenuation[ch]) / 100.0f);
+    float offsetValue     = attenuatedValue + (CVInputOffset[ch] / 100.0f * MAXDAC);
+    CVValue = constrain(offsetValue, 0, MAXDAC);
+
+    // CRITICAL SECTION: protect parameter updates that affect timing
+    noInterrupts();
+
+    switch (cvTarget) {
+    case CVTarget::None:
+        break;
+    case CVTarget::StartStop:
+        if (CVValue > MAXDAC / 2) {
+            SetMasterState(true);
+        } else {
+            SetMasterState(false);
+        }
+        break;
+    case CVTarget::Reset:
+        if (CVValue > MAXDAC / 2 && !lastResetState) {
+            tickCounter        = 0;
+            externalTickCounter = 0;
+            lastResetState     = true;
+        } else if (CVValue < MAXDAC / 2) {
+            lastResetState = false;
+        }
+        break;
+    case CVTarget::SetBPM:
+        UpdateBPM(map(CVValue, 0, MAXDAC, minBPM, maxBPM));
+        break;
+    case CVTarget::Div1:
+        outputs[0].SetDivider(map(CVValue, 0, MAXDAC, 0, outputs[0].GetDividerAmounts()));
+        break;
+    case CVTarget::Div2:
+        outputs[1].SetDivider(map(CVValue, 0, MAXDAC, 0, outputs[1].GetDividerAmounts()));
+        break;
+    case CVTarget::Div3:
+        outputs[2].SetDivider(map(CVValue, 0, MAXDAC, 0, outputs[2].GetDividerAmounts()));
+        break;
+    case CVTarget::Div4:
+        outputs[3].SetDivider(map(CVValue, 0, MAXDAC, 0, outputs[3].GetDividerAmounts()));
+        break;
+    case CVTarget::Output1Prob:
+        outputs[0].SetPulseProbability(map(CVValue, 0, MAXDAC, 1, 100));
+        break;
+    case CVTarget::Output2Prob:
+        outputs[1].SetPulseProbability(map(CVValue, 0, MAXDAC, 1, 100));
+        break;
+    case CVTarget::Output3Prob:
+        outputs[2].SetPulseProbability(map(CVValue, 0, MAXDAC, 1, 100));
+        break;
+    case CVTarget::Output4Prob:
+        outputs[3].SetPulseProbability(map(CVValue, 0, MAXDAC, 1, 100));
+        break;
+    case CVTarget::Swing1Amount:
+        outputs[0].SetSwingAmount(map(CVValue, 0, MAXDAC, 0, outputs[0].GetSwingAmounts()));
+        break;
+    case CVTarget::Swing1Every:
+        outputs[0].SetSwingEvery(map(CVValue, 0, MAXDAC, 1, outputs[0].GetSwingEveryAmounts()));
+        break;
+    case CVTarget::Swing2Amount:
+        outputs[1].SetSwingAmount(map(CVValue, 0, MAXDAC, 0, outputs[1].GetSwingAmounts()));
+        break;
+    case CVTarget::Swing2Every:
+        outputs[1].SetSwingEvery(map(CVValue, 0, MAXDAC, 1, outputs[1].GetSwingEveryAmounts()));
+        break;
+    case CVTarget::Swing3Amount:
+        outputs[2].SetSwingAmount(map(CVValue, 0, MAXDAC, 0, outputs[2].GetSwingAmounts()));
+        break;
+    case CVTarget::Swing3Every:
+        outputs[2].SetSwingEvery(map(CVValue, 0, MAXDAC, 1, outputs[2].GetSwingEveryAmounts()));
+        break;
+    case CVTarget::Swing4Amount:
+        outputs[3].SetSwingAmount(map(CVValue, 0, MAXDAC, 0, outputs[3].GetSwingAmounts()));
+        break;
+    case CVTarget::Swing4Every:
+        outputs[3].SetSwingEvery(map(CVValue, 0, MAXDAC, 1, outputs[3].GetSwingEveryAmounts()));
+        break;
+    case CVTarget::Output3Offset:
+        outputs[2].SetOffset(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output4Offset:
+        outputs[3].SetOffset(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output3Level:
+        outputs[2].SetLevel(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output4Level:
+        outputs[3].SetLevel(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output3Waveform:
+        outputs[2].SetWaveformType(static_cast<WaveformType>(map(CVValue, 0, MAXDAC, 0, WaveformTypeLength)));
+        break;
+    case CVTarget::Output4Waveform:
+        outputs[3].SetWaveformType(static_cast<WaveformType>(map(CVValue, 0, MAXDAC, 0, WaveformTypeLength)));
+        break;
+    case CVTarget::Output1Duty:
+        outputs[0].SetDutyCycle(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output2Duty:
+        outputs[1].SetDutyCycle(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output3Duty:
+        outputs[2].SetDutyCycle(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Output4Duty:
+        outputs[3].SetDutyCycle(map(CVValue, 0, MAXDAC, 0, 100));
+        break;
+    case CVTarget::Envelope1:
+        outputs[2].SetExternalTrigger(CVValue > MAXDAC / 2);
+        break;
+    case CVTarget::Envelope2:
+        outputs[3].SetExternalTrigger(CVValue > MAXDAC / 2);
+        break;
+    case CVTarget::Output3:
+        outputs[2].SetCVValue(CVValue);
+        break;
+    case CVTarget::Output4:
+        outputs[3].SetCVValue(CVValue);
+        break;
+    }
+
+    interrupts();
+}
